move modules crank.js -> core.js, move mod.js -> crank.js
This is done mainly to appease esm.sh, which seems to get confused when you have
a module which is named the same as the package.
