--- src/mod.ts
@@ -1,2 +1,0 @@
-export * from "./crank.js";
-export {jsx} from "./tags.js";
